NewsFeed (API request + basic UI setup)
Now have the news feed up and running on one of the tab views. It is able to fetch any revelent news about the Mars missions and display them to the user.

The UI is also setup to display the news but it's nothing fancy and still doesn't have any images on it yet.

Will also need to look into limiting the languages that are pulled from the API