API Issue + Infinite Scrolling established (I believe)
Getting the error: nw_protocol_get_quic_image_block_invoke dlopen libquic failed when trying to run the app currently. This has stopped the API from running and may be Pod related by the looks of things.

Everything is currently in place to have infinite scrolling (assuming the source allows it). I've just tried the API in another app and got a similar message to appears to be an XCode issue?

URL emebedded in the code displays perfectly okay in browser and request are being registered on the API's page