Image from API + Caching
Now have the images from the API being displayed on the UI alongside the title and author. I've created a new model for the UIImages and adjusted the NewsView to ensure that the pictures show

I have also established a cache for the images to improve the performance.The images are now held in the cache after being loaded for the first time and then drawn from the cache if they need to be displayed again. This allows the scroll on the UI to run a little nicer

Will look into using the NASA API next to get images from the Rovers