PhotoView updated + Code Tidy
Have changed how the JSON data is displayed a little but still far off where I'd want it to be. Font needs to be jazzed up and working on the alignment too. Need to change how the date is displayed too

Also just a general code tidy from what I've put together and added comments as well.