PhotoView Tidy
Have updated the PhotoView somewhat to make it look a little more appealing. Have the Mars APIs now displaying in horizontal ScrollViews as it's way more presentable that way (far easier to work than a collectionView as well).

Have moved the Photo Of The Day (POTD) to a subview that is then called in the PhotoView. Modifiers have been placed on this subview and the image displays as it should.

Will next work on making the two Mars API dates of some significance, then updating the UI to give it some life